+++ plugwise/util.py
@@ -1,11 +1,9 @@
 """Plugwise protocol helpers."""
 from __future__ import annotations
 
-import datetime
 import re
 
 from .constants import (
-    ARBITRARY_DATE,
     ELECTRIC_POTENTIAL_VOLT,
     ENERGY_KILO_WATT_HOUR,
     HW_MODELS,
@@ -49,28 +47,6 @@ def format_measure(measure: str, unit: str) -> float | int:
                 result = int(round(float_measure))
 
     return result
-
-
-def in_between(
-    today: int,
-    day_0: int,
-    day_1: int,
-    now: datetime.time,
-    time_0: datetime.time,
-    time_1: datetime.time,
-) -> bool:
-    """Determine timing for schedules."""
-    time_now = datetime.timedelta(days=today, hours=now.hour, minutes=now.minute)
-    time_start = datetime.timedelta(
-        days=day_0, hours=time_0.hour, minutes=time_0.minute
-    )
-    time_end = datetime.timedelta(days=day_1, hours=time_1.hour, minutes=time_1.minute)
-
-    now_point = ARBITRARY_DATE + time_now
-    start_point = ARBITRARY_DATE + time_start
-    end_point = ARBITRARY_DATE + time_end
-
-    return start_point <= now_point <= end_point
 
 
 # NOTE: this function version_to_model is shared between Smile and USB