+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## Ongoing
+## v0.33.2 Bugfix for HA climate
 
+- Remove last_used (schedule) from output.
 - Bugfix for [Core Issue #102204](https://github.com/home-assistant/core/issues/102204)
 - Add item-count to output
 - Support python 3.12